Migrated shared build logic from buildSrc to composite builds
The sharing code is rewritten as convention plugins and placed in the separate build.

Two plugins are allocated - for setting up the publication and for setting up the Kotlin multiplatform project.

Resolves #142
