+++ build-logic/src/main/kotlin/kotlinx/io/conventions/multiplatform-lib-conventions.gradle.kts
@@ -3,10 +3,13 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENCE file.
  */
 
-import org.gradle.api.*
-import org.gradle.kotlin.dsl.*
+import org.gradle.kotlin.dsl.kotlin
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
-import org.jetbrains.kotlin.gradle.plugin.*
+import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
+
+plugins {
+    kotlin("multiplatform")
+}
 
 fun KotlinMultiplatformExtension.configureNativePlatforms() {
     iosX64()
@@ -23,7 +26,7 @@ fun KotlinMultiplatformExtension.configureNativePlatforms() {
     linuxArm64()
     androidNativeArm32()
     androidNativeArm64()
-        androidNativeX64()
+    androidNativeX64()
     androidNativeX86()
     // Required to generate tests tasks: https://youtrack.jetbrains.com/issue/KT-26547
     linuxX64()
@@ -93,5 +96,16 @@ fun NamedDomainObjectContainer<KotlinSourceSet>.createSourceSet(
     }
 
     return result
+}
+
+kotlin {
+    configureNativePlatforms()
+    sourceSets {
+        val commonMain by getting
+        val commonTest by getting
+
+        createSourceSet("nativeMain", parent = commonMain, children = nativeTargets)
+        createSourceSet("nativeTest", parent = commonTest, children = nativeTargets)
+    }
 }
