+++ build.gradle.kts
@@ -9,19 +9,8 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinNativeCompile
 plugins {
     alias(libs.plugins.bcv)
     alias(libs.plugins.dokka)
-    `maven-publish`
-    signing
-}
 
-buildscript {
-    dependencies {
-        classpath(libs.kotlin.gradle.plugin)
-    }
-
-    repositories {
-        mavenCentral()
-        gradlePluginPortal()
-    }
+    id("publish-conventions") apply false
 }
 
 allprojects {
@@ -31,9 +20,6 @@ allprojects {
     }
 }
 
-apply(plugin = "maven-publish")
-apply(plugin = "signing")
-
 subprojects {
     if (name.contains("benchmark")) {
         return@subprojects
@@ -41,22 +27,6 @@ subprojects {
 
     repositories {
         mavenCentral()
-    }
-
-    apply(plugin = "maven-publish")
-    apply(plugin = "signing")
-
-    publishing {
-        repositories {
-            configureMavenPublication(project)
-        }
-
-        val javadocJar = project.configureEmptyJavadocArtifact()
-        publications.withType(MavenPublication::class).all {
-            pom.configureMavenCentralMetadata(project)
-            signPublicationIfKeyPresent(project, this)
-            artifact(javadocJar)
-        }
     }
 }
 
@@ -72,6 +42,8 @@ subprojects {
             allWarningsAsErrors = true
         }
     }
+
+    apply(plugin = "publish-conventions")
 }
 
 apiValidation {
