+++ bytestring/build.gradle.kts
@@ -2,7 +2,7 @@ import org.jetbrains.dokka.gradle.DokkaTaskPartial
 import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
 
 plugins {
-    kotlin("multiplatform")
+    id("multiplatform-lib-conventions")
     alias(libs.plugins.kover)
     alias(libs.plugins.dokka)
 }
@@ -36,7 +36,6 @@ kotlin {
         }
     }
 
-    configureNativePlatforms()
     sourceSets {
         val commonMain by getting
         val commonTest by getting {
@@ -44,9 +43,6 @@ kotlin {
                 implementation(kotlin("test"))
             }
         }
-
-        createSourceSet("nativeMain", parent = commonMain, children = nativeTargets)
-        createSourceSet("nativeTest", parent = commonTest, children = nativeTargets)
     }
 
     explicitApi()
