+++ settings.gradle.kts
@@ -4,6 +4,8 @@
  */
 
 pluginManagement {
+    includeBuild("build-logic")
+
     repositories {
         mavenCentral()
         google()
